[feature]Homepage and AddProjectpage frame done and fix the problem of the img doesn't show.
For the [fix] problem,because the angular compiler cant read the inner
folder of Components , so when setting the src path inside it and it
will show "image 404 not found" in Browser.
The only asset folder angular can access is behind the src ,so the path
of img needs to be set in the src/assets.